Use API to determine permissions
